refactor core rule parser
